Tests/LibWeb/Text/expected/geolocation-watch-emulated.txt

update 1: 37.7647658, -122.4345892
update 2: 1.25, 2.5
